refactor for better task arrangement and readability
